# hospital-service/docs/swagger.yaml
basePath: /
definitions:
  appointment.Appointment:
    properties:
      appointment_code:
        description: Mã lịch hẹn duy nhất
        type: string
      appointment_id:
        type: string
      book_user_id:
        type: string
      checked_in_at:
        description: Thời gian bệnh nhân check-in (nếu có)
        type: string
      created_at:
        type: string
      doctor:
        $ref: '#/definitions/doctor.Doctor'
      doctor_id:
        description: ID bác sĩ
        type: string
      hospital:
        $ref: '#/definitions/hospital.Hospital'
      hospital_id:
        description: ID bệnh viện
        type: string
      notes:
        description: Ghi chú thêm từ bệnh nhân
        type: string
      patient:
        allOf:
        - $ref: '#/definitions/patient.Patient'
        description: Quan hệ
      patient_id:
        description: ID bệnh nhân
        type: string
      related_record_id:
        description: nếu là tái khám
        type: string
      service_name:
        description: Cho phép empty string làm default
        type: string
      status:
        $ref: '#/definitions/enums.AppointmentStatus'
      time_slots:
        items:
          $ref: '#/definitions/appointment.TimeSlot'
        type: array
      updated_at:
        type: string
    type: object
  appointment.TimeSlot:
    properties:
      appointment:
        $ref: '#/definitions/appointment.Appointment'
      appointment_id:
        type: string
      capacity:
        type: integer
      created_at:
        type: string
      doctor:
        $ref: '#/definitions/doctor.Doctor'
      doctor_id:
        type: string
      end_time:
        type: string
      slot_id:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
    type: object
  appointmenthandler.UpdateAppointmentStatusRequest:
    properties:
      status:
        $ref: '#/definitions/enums.AppointmentStatus'
    required:
    - status
    type: object
...
  doctor.Doctor:
    properties:
      created_at:
        type: string
      doctor_id:
        type: string
      email:
        type: string
      full_name:
        type: string
      hospital_id:
        type: string
      image:
        type: string
      phone:
        type: string
      slug:
        type: string
      specialty:
        $ref: '#/definitions/enums.Specialty'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  enums.AppointmentStatus:
    enum:
    - PENDING
    - CONFIRMED
    - COMPLETED
    - CANCELED
    - PENDING_ONLINE
    - CONFIRMED_ONLINE
    - COMPLETED_ONLINE
    type: string
    x-enum-varnames:
    - Pending
    - Confirmed
    - Completed
    - Canceled
    - PendingOnline
    - ConfirmedOnline
    - CompletedOnline
...
  enums.Gender:
    enum:
    - male
    - female
    type: string
    x-enum-varnames:
    - Male
    - Female
...
  enums.Specialty:
    enum:
    - ophthalmology
    - retina
    - cornea
    - glaucoma
    - refractive_surgery
    - pediatric_ophthalm
    - schedule_manager
    type: string
    x-enum-comments:
      SpecialtyCornea: Chuyên khoa giác mạc
      SpecialtyGlaucoma: Cườm nước
      SpecialtyOphthalmology: Nhãn khoa tổng quát
      SpecialtyPediatricOphthalm: Nhãn khoa nhi
      SpecialtyRefractiveSurgery: Phẫu thuật khúc xạ (LASIK)
      SpecialtyRetina: Chuyên khoa võng mạc
      SpecialtyScheduleManager: Quản lý lịch bác sĩ
    x-enum-descriptions:
    - Nhãn khoa tổng quát
    - Chuyên khoa võng mạc
    - Chuyên khoa giác mạc
    - Cườm nước
    - Phẫu thuật khúc xạ (LASIK)
    - Nhãn khoa nhi
    - Quản lý lịch bác sĩ
    x-enum-varnames:
    - SpecialtyOphthalmology
    - SpecialtyRetina
    - SpecialtyCornea
    - SpecialtyGlaucoma
    - SpecialtyRefractiveSurgery
    - SpecialtyPediatricOphthalm
    - SpecialtyScheduleManager
  hospital.Hospital:
    properties:
      address:
        type: string
      city:
        type: string
      created_at:
        type: string
      doctors:
        items:
          $ref: '#/definitions/doctor.Doctor'
        type: array
      email:
        type: string
      hospital_id:
        type: string
      image:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      phone:
        type: string
      slug:
        type: string
      updated_at:
        type: string
      url_map:
        type: string
      ward:
        type: string
    type: object
...
  patient.Patient:
    properties:
      address:
        type: string
      created_at:
        type: string
      dob:
        type: string
      email:
        type: string
      full_name:
        type: string
      gender:
        $ref: '#/definitions/enums.Gender'
      image:
        type: string
      patient_id:
        type: string
      phone:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8084
info:
  contact: {}
  title: Auth Service API
  version: "1.0"
paths:
...
  /appointments:
    get:
      description: Retrieve a list of all appointments in the system with optional
        filters
      parameters:
      - description: Filter by patient name (partial match)
        in: query
        name: patient_name
        type: string
      - description: Filter by appointment status (PENDING/CONFIRMED/COMPLETED/CANCELLED)
        in: query
        name: status
        type: string
      - description: Filter by doctor ID (exact match)
        in: query
        name: doctor_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/appointment.Appointment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all appointments
      tags:
      - Appointments
  /appointments/{appointment_id}:
    delete:
      description: Remove an appointment by its ID
      parameters:
      - description: Appointment ID
        in: path
        name: appointment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an appointment
      tags:
      - Appointments
    get:
      description: Retrieve an appointment record using its ID
      parameters:
      - description: Appointment ID
        in: path
        name: appointment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/appointment.Appointment'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get appointment by ID
      tags:
      - Appointments
  /appointments/{appointment_id}/detail:
    put:
      consumes:
      - application/json
      description: Update the details of an existing appointment (date, notes, etc.)
      parameters:
      - description: Appointment ID
        in: path
        name: appointment_id
        required: true
        type: string
      - description: Updated appointment data
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/appointment.Appointment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update appointment details
      tags:
      - Appointments
  /appointments/{appointment_id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an appointment (PENDING, CONFIRMED, COMPLETED,
        CANCELED)
      parameters:
      - description: Appointment ID
        in: path
        name: appointment_id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/appointmenthandler.UpdateAppointmentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update appointment status
      tags:
      - Appointments
  /appointments/doctor/{doctor_id}:
    get:
      description: Retrieve all appointments assigned to a specific doctor
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/appointment.Appointment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get appointments by doctor ID
      tags:
      - Appointments
  /appointments/online:
    get:
      description: Get all online appointments (status = PendingOnline) by bookUserID
        or doctorID
      parameters:
      - description: Book user ID
        in: query
        name: book_user_id
        type: string
      - description: Doctor ID
        in: query
        name: doctor_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/appointment.Appointment'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get online appointments
      tags:
      - Appointments
  /appointments/patient/{patient_id}:
    get:
      description: Retrieve all appointments belonging to a specific patient
      parameters:
      - description: Patient ID
        in: path
        name: patient_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/appointment.Appointment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get appointments by patient ID
      tags:
      - Appointments
...
swagger: "2.0"